+++ make.settings
@@ -5,6 +5,7 @@ PLATFORM_CC                     = gcc
 PLATFORM_AR                     = ar
 PLATFORM_OS                     = linux
 PLATFORM_SSL                    = mbedtls
+PLATFORM_MODULE		            = m26
 
 # Uncomment below and specify PATH to your toolchain when cross-compile SDK
 #
@@ -21,4 +22,6 @@ FEATURE_FILE_UPLOAD_ENABLED             = y     # 是否打开文件上传
 
 FEATURE_AUTH_MODE_DYNAMIC               = y     # 是否使能设备动态注册
 FEATURE_SUPPORT_TLS                     = y     # 是否打开TLS支持
+FEATURE_SUPPORT_AT_CMD                  = n     # 是否打开AT命令支持
+
 FEATURE_SDK_TESTS_ENABLED               = n     # 是否打开SDK测试用例编译
