+++ src/at/class/ec20/ec20.c
@@ -15,45 +15,23 @@
 
 #include <stdio.h>
 #include <string.h>
-#include "stm32f7xx_hal.h"
 #include "at_inf.h"
 #include "utils_net.h"
 #include "at_ringbuff.h"
 #include "at_client.h"
 #include "uiot_import.h"
+#include "at_utils.h"
 
 #define EC20_MAX_TCP_LINK        3
 
-extern UART_HandleTypeDef huart2;
-static UART_HandleTypeDef *pAtUart = &huart2;
 extern sRingbuff g_ring_buff;    
 extern sRingbuff g_ring_tcp_buff[3];    
 int ec20_link[EC20_MAX_TCP_LINK] = {0};
 int ec20_recv_len[EC20_MAX_TCP_LINK] = {0};
 
 extern int last_tcp_link;
-int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_ size_t len)
-{
-    int ret = 0;
-
-    if(pAtUart->RxState == HAL_UART_STATE_BUSY_RX)
-    {
-        HAL_SleepMs(10);
-    }
-
-    ret = ring_buff_pop_data(&(g_ring_tcp_buff[pNetwork->handle-1]), buffer, len);
-
-    return ret;
-}
-
-int HAL_AT_Write(_IN_ unsigned char *buffer, _IN_ size_t len)
-{   
-    return HAL_UART_Transmit_IT(pAtUart, buffer, len);
-}
-
 int HAL_AT_Read_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
 {
-    at_client_t client = at_client_get();
     at_response_t resp = NULL;
     
     resp = at_create_resp(256, 0, CMD_TIMEOUT_MS);
@@ -104,7 +82,7 @@ int HAL_AT_Write_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer,
     resp->custom_flag = true;
     at_exec_cmd(resp, at_command, 0, "AT+QISEND=%d,%d\r\n", pNetwork->handle-1, len); 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_data, len, buffer); 
+    ret = at_exec_cmd(resp, at_data, len, (const char *)buffer); 
     HAL_SleepMs(100);
 
     at_delete_resp(resp);
@@ -388,7 +366,7 @@ static int urc_qird_recv_func(const char *data, uint32_t size)
             for(loop = recv_data_num; loop > 0; loop--)
             {
                 ret = at_client_getchar(client, &temp_char, GET_RECEIVE_TIMEOUT_MS);
-                ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), &temp_char, 1);
+                ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), (uint8_t *)&temp_char, 1);
                 if(SUCCESS_RET != ret)
                 {
                     LOG_ERROR("copy data to tcp buff fail\n");
@@ -436,9 +414,6 @@ static int ec20_init()
         LOG_ERROR("No memory for response object!");
         return FAILURE_RET;
     }
-
-    /* 配置串口接收buf的存储位置 */    
-    HAL_UART_Receive_IT(pAtUart, g_ring_buff.buffer, 1);
 
     //延时等待模块启动
     HAL_SleepMs(5000);
@@ -595,10 +570,9 @@ end:
     return ret;
 }
 
-int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
+int HAL_AT_TCP_Connect(_IN_ utils_network_pt pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
 {
     int ret = 0;
-    utils_network_pt pNet = (utils_network_pt)pNetwork;
     at_response_t resp = NULL;
     int link_num = 0;
     at_client_t p_client = at_client_get();
@@ -645,8 +619,7 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
 
     /* 建立TCP链接 */
     resp->custom_flag = true;
-    //ret = at_exec_cmd(resp, at_command, 0,  "AT+QIOPEN=1,%d,\"TCP\",\"%s\",%d,0,1\r\n", link_num, pNet->pHostAddress, pNet->port);
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIOPEN=1,%d,\"TCP\",\"%s\",%d,0,0\r\n", link_num, pNet->pHostAddress, pNet->port);
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIOPEN=1,%d,\"TCP\",\"%s\",%d,0,0\r\n", link_num, pNetwork->pHostAddress, pNetwork->port);
     
     if(SUCCESS_RET != ret)
     {
@@ -657,7 +630,7 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
     {
         ec20_link[link_num] = eCONNECTED;
         /* handle can't be zero */
-        pNet->handle = link_num + 1;
+        pNetwork->handle = link_num + 1;
     }
 end:
     at_delete_resp(resp);
