+++ src/at/class/m26/m26.c
@@ -12,25 +12,20 @@
 * express or implied. See the License for the specific language governing
 * permissions and limitations under the License.
 */
-
 #include <stdio.h>
 #include <string.h>
-#include "stm32f7xx_hal.h"
 #include "at_inf.h"
 #include "utils_net.h"
 #include "at_ringbuff.h"
 #include "at_client.h"
 #include "uiot_import.h"
 
-#define SIM800C_MAX_TCP_LINK        3
-
-extern UART_HandleTypeDef huart2;
-static UART_HandleTypeDef *pAtUart = &huart2;
+#define M26_MAX_TCP_LINK        3
 extern sRingbuff g_ring_buff;    
 extern sRingbuff g_ring_tcp_buff[3];    
-int sim800c_link[SIM800C_MAX_TCP_LINK] = {0};
+int m26_link[M26_MAX_TCP_LINK] = {0};
 extern int last_tcp_link;
-int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_ size_t len)
+int HAL_AT_Read_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
 {
     int ret = 0;
     at_client_t client = at_client_get();
@@ -41,11 +36,6 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
     //char temp_tcp = 0;
     int rec_num = 0;
     int link_num = 0;
-
-    if(pAtUart->RxState == HAL_UART_STATE_BUSY_RX)
-    {
-        HAL_SleepMs(10);
-    }
 
     /* clear \r\n */
     ret = at_client_getchar(client, &last_char, GET_RECEIVE_TIMEOUT_MS);
@@ -76,7 +66,7 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
                 if(loop == 8)
                 {   
                     /* divide multi tcp link */
-                    if(0 != strncmp(temp_string, "RECEIVE,", 8))
+                    if(0 != strncmp(temp_string, "RECEIVE:", 8))
                     {
                         client->pRingBuff->readpoint = temp_read_point;
                         break;
@@ -85,12 +75,12 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
                 loop++;
             }while(temp_string[loop-1] != '\n');
             temp_string[loop] = '\0';
-            if(2 == sscanf(temp_string,"RECEIVE,%d,%d:\r\n",&link_num,&rec_num))
+            if(2 == sscanf(temp_string,"RECEIVE: %d, %d:\r\n",&link_num,&rec_num))
             {
                 for(loop = rec_num; loop > 0; loop--)
                 {
                     ret = at_client_getchar(client, &temp_char, GET_RECEIVE_TIMEOUT_MS);
-                    ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), &temp_char, 1);
+                    ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), (uint8_t *)&temp_char, 1);
                     if(SUCCESS_RET != ret)
                     {
                         LOG_ERROR("copy data to tcp buff fail\n");
@@ -106,21 +96,11 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
     }
 
 
-    ret = ring_buff_pop_data(&(g_ring_tcp_buff[pNetwork->handle-1]), buffer, len);
+    ret = HAL_AT_Read(pNetwork, buffer, len);
 
     return ret;
 }
 
-
-int HAL_AT_Write(_IN_ unsigned char *buffer, _IN_ size_t len)
-{   
-    return HAL_UART_Transmit_IT(pAtUart, buffer, len);
-}
-
-int HAL_AT_Read_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
-{
-    return HAL_AT_Read(pNetwork, buffer, len);
-}
 
 int HAL_AT_Write_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
 {
@@ -140,9 +120,9 @@ int HAL_AT_Write_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer,
     }
 
     resp->custom_flag = true;
-    at_exec_cmd(resp, at_command, 0, "AT+CIPSEND=%d,%d\r\n", pNetwork->handle-1, len); 
+    at_exec_cmd(resp, at_command, 0, "AT+QISEND=%d,%d\r\n", pNetwork->handle-1, len); 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_data, len, buffer); 
+    ret = at_exec_cmd(resp, at_data, len, (const char *)buffer); 
     HAL_SleepMs(100);
 
     at_delete_resp(resp);
@@ -162,7 +142,7 @@ int HAL_AT_TCP_Disconnect(utils_network_pt pNetwork)
     int ret = SUCCESS_RET;
     
     /* 断开无线链接 */
-    if(sim800c_link[pNetwork->handle-1] == eCONNECTED)
+    if(m26_link[pNetwork->handle-1] == eCONNECTED)
     {
         at_response_t resp = NULL;
         
@@ -174,21 +154,19 @@ int HAL_AT_TCP_Disconnect(utils_network_pt pNetwork)
         }
     
         resp->custom_flag = false;
-        ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPCLOSE=%d\r",pNetwork->handle-1);
+        ret = at_exec_cmd(resp, at_command, 0,  "AT+QICLOSE=%d\r",pNetwork->handle-1);
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("close TCP link fail!\n");
         }
         else
         {
-            sim800c_link[pNetwork->handle-1] = eDISCONNECTED;
+            m26_link[pNetwork->handle-1] = eDISCONNECTED;
         }
         at_delete_resp(resp);
     }
     return ret;
 }
-
-char ip_addr[20] = {0};
 
 static int urc_cpin_recv_judge(const char *data, uint32_t size)
 {
@@ -260,7 +238,7 @@ static int urc_creg_recv_func(const char *data, uint32_t size)
     int stat = 0;
     if(2 == sscanf(data,"+CREG: %d,%d", &n, &stat))
     {
-        if(1 == stat)
+        if((1 == stat) || (5 == stat))
         {
             return SUCCESS_RET;
         }
@@ -270,10 +248,9 @@ static int urc_creg_recv_func(const char *data, uint32_t size)
 }
 
 
-static int urc_cgatt_recv_judge(const char *data, uint32_t size)
+static int urc_cgreg_recv_judge(const char *data, uint32_t size)
 {
-    if((data[size-1] == '\n')
-        && (NULL != strstr(data, "+CGATT")))
+    if(NULL != strstr(data, "+CGREG"))
     {
         return SUCCESS_RET;
     }
@@ -283,16 +260,14 @@ static int urc_cgatt_recv_judge(const char *data, uint32_t size)
     }
 }
 
-static int urc_cgatt_recv_func(const char *data, uint32_t size)
+static int urc_cgreg_recv_func(const char *data, uint32_t size)
 {
-    at_client_t client = at_client_get();
-    int result = 0;
-    if(1 == sscanf(data,"+CGATT: %d",&result))
+    int n = 0;
+    int stat = 0;
+    if(2 == sscanf(data,"+CGREG: %d,%d",&n, &stat))
     {
-        if(1 == result)
+        if((1 == stat) || (5 == stat))
         {
-            at_recv_readline(client);
-            at_recv_readline(client);
             return SUCCESS_RET;
         }
     }
@@ -306,11 +281,52 @@ static int urc_common_recv_func(const char *data, uint32_t size)
     return SUCCESS_RET;
 }
 
+static int urc_qimode_recv_judge(const char *data, uint32_t size)
+{
+    if(NULL != strstr(data, "+QIMODE:"))
+    {
+        return SUCCESS_RET;
+    }
+    else
+    {
+        return FAILURE_RET;
+    }
+}
+
+static int urc_qimode_recv_func(const char *data, uint32_t size)
+{
+    at_client_t client = at_client_get();
+    int n = 0;
+    if(1 == sscanf(data,"+QIMODE: %d\r\n",&n))
+    {
+        if(0 == n)
+        {        
+            at_recv_readline(client);    
+            at_recv_readline(client);
+            return SUCCESS_RET;
+        }
+    }
+    
+    return FAILURE_RET;
+}
+
+static int urc_qideact_recv_judge(const char *data, uint32_t size)
+{
+    if(NULL != strstr(data, "DEACT OK"))
+    {
+        return SUCCESS_RET;
+    }
+    else
+    {
+        return FAILURE_RET;
+    }
+}
+
 static int urc_ip_recv_judge(const char *data, uint32_t size)
 {
     int num1,num2,num3,num4;
     if((data[size-1] == '\n')
-        && (4 == sscanf(data,"%d.%d.%d.%d",&num1,&num2,&num3,&num4)))
+        && (4 == sscanf(data,"%d.%d.%d.%d\r\n",&num1,&num2,&num3,&num4)))
     {   
         if(0<=num1 && num1<=255
          && 0<=num2 && num2<=255
@@ -320,15 +336,21 @@ static int urc_ip_recv_judge(const char *data, uint32_t size)
             return SUCCESS_RET;
         }
     }
-    else
-    {
-        return FAILURE_RET;
-    }
+        
+    return FAILURE_RET;
 }
 
+char dev_ip_addr[20] = {0};
 static int urc_ip_recv_func(const char *data, uint32_t size)
 {
-    strncpy(ip_addr, data, size);
+    strncpy(dev_ip_addr, data, size-2);
+    return SUCCESS_RET;
+}
+
+char domain_ip_addr[20] = {0};
+static int urc_domain_ip_recv_func(const char *data, uint32_t size)
+{
+    strncpy(domain_ip_addr, data, size-2);
     return SUCCESS_RET;
 }
 
@@ -368,21 +390,24 @@ static int urc_close_recv_judge(const char *data, uint32_t size)
     }
 }
 
-
 at_custom custom_table[] = {
     {"AT+CPIN?", 12, urc_cpin_recv_judge, urc_cpin_recv_func},
     {"AT+CSQ", 10, urc_csq_recv_judge, urc_csq_recv_func},
     {"AT+CREG?", 10, urc_creg_recv_judge, urc_creg_recv_func},
-    {"AT+CGATT?", 9, urc_cgatt_recv_judge, urc_cgatt_recv_func},
-    {"AT+CIFSR", 10, urc_ip_recv_judge, urc_ip_recv_func},
-    {"AT+CIPSTART", 14, urc_tcp_start_judge, urc_common_recv_func},
-    {"AT+CIPSEND", 1, urc_send_recv_judge, urc_common_recv_func},
-    {"AT+CIPCLOSE", 12, urc_close_recv_judge, urc_common_recv_func},
+    {"AT+CGREG?", 11, urc_cgreg_recv_judge, urc_cgreg_recv_func},
+    {"AT+QIMODE?", 12, urc_qimode_recv_judge, urc_qimode_recv_func},
+    {"AT+QIDEACT?", 9, urc_qideact_recv_judge, urc_common_recv_func},
+    {"AT+QILOCIP", 13, urc_ip_recv_judge, urc_ip_recv_func},
+    {"AT+QIDNSGIP", 13, urc_ip_recv_judge, urc_domain_ip_recv_func},
+    {"AT+QIDEACT", 10, urc_qideact_recv_judge, urc_common_recv_func},
+    {"AT+QIOPEN", 14, urc_tcp_start_judge, urc_common_recv_func},
+    {"AT+QISEND", 1, urc_send_recv_judge, urc_common_recv_func},
+    {"AT+QICLOSE", 12, urc_close_recv_judge, urc_common_recv_func},
 };
 
 int custom_table_num = sizeof(custom_table) / sizeof(custom_table[0]);
 
-static int sim800c_init()
+static int m26_init()
 {
     int ret = 0;
     at_response_t resp = NULL;
@@ -395,15 +420,12 @@ static int sim800c_init()
         return FAILURE_RET;
     }
 
-    /* 配置串口接收buf的存储位置 */    
-    HAL_UART_Receive_IT(pAtUart, g_ring_buff.buffer, 1);
-
     //延时等待模块启动
-    HAL_SleepMs(8000);
+    HAL_SleepMs(5000);
 
     /* 去掉串口回显 */
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0, "ATE0\r");
+    ret = at_exec_cmd(resp, at_command, 0, "ATE0\r\n");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("ATE0 set fail!\n");
@@ -419,7 +441,7 @@ static int sim800c_init()
         
         /* 检查SIM卡状态 */
         resp->custom_flag = true;
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CPIN?\r");  
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CPIN?\r\n");  
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("please check SIM card status!\n");
@@ -428,7 +450,7 @@ static int sim800c_init()
         
         /* 检查网络强度 */  
         resp->custom_flag = true;
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CSQ\r");  
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CSQ\r\n");  
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("bad signal!\n");
@@ -437,7 +459,7 @@ static int sim800c_init()
         
         /* 检查网络注册状态 */   
         resp->custom_flag = true;   
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CREG?\r"); 
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CREG?\r\n"); 
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("not network regist!\n");
@@ -446,11 +468,20 @@ static int sim800c_init()
 
         /*检查GPRS附着状态*/
         resp->custom_flag = true;   
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CGATT?\r"); 
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CGREG?\r\n"); 
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("not Attach GPRS!\n");
             continue;
+        }
+
+        /* check TCPIP mode is set */
+        resp->custom_flag = true;
+        ret = at_exec_cmd(resp, at_command, 0,  "AT+QIMODE?\r\n");
+        if(SUCCESS_RET != ret)
+        {
+            LOG_ERROR("check TCPIP mode fail!\n");
+            goto end;
         }
 
         if(SUCCESS_RET == ret)
@@ -466,27 +497,54 @@ static int sim800c_init()
         goto end;
     }
 
+    /* 设置前台配置 */
+    resp->custom_flag = false;
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIFGCNT=0\r\n");
+    if(SUCCESS_RET != ret)
+    {
+        LOG_ERROR("set foreground fail!\n");
+        goto end;
+    }
+    
+    /* set the apn */
+    resp->custom_flag = false;
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QICSGP=1, \"CMNET\"\r\n");
+    if(SUCCESS_RET != ret)
+    {
+        LOG_ERROR("set apn fail!\n");
+        goto end;
+    }
+
+    /* 关闭 GPRS/CSD PDP 场景 */    
+    resp->custom_flag = true;
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIDEACT\r\n");
+    if(SUCCESS_RET != ret)
+    {
+        LOG_ERROR("shut down GPRS/CSD PDP fail!\n");
+        goto end;
+    }
+
     /* 设置多链路模式 */ 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPMUX=1\r");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIMUX=1\r\n");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("set multi link mode fail!\n");
         goto end;
     }
     
-    /* 设置APN */    
+    /* 启动任务 */    
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CSTT=\"3GNET\"\r");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIREGAPP\r\n");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("set APN fail!\n");
         goto end;
     }
     
-    /* 建立无线链接 */ 
+    /* 设备默认响应超时时间为150s，设置为20s */ 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIICR\r");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIACT\r\n");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("build wireless link fail!\n");
@@ -495,19 +553,10 @@ static int sim800c_init()
     
     /* 获取本地IP地址 */  
     resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIFSR\r");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QILOCIP\r\n");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("fetch local IP address fail!\n");
-        goto end;
-    }
-
-    /* 快速发送模式 */  
-    resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPQSEND=1\r");
-    if(SUCCESS_RET != ret)
-    {
-        LOG_ERROR("fast send mode fail!\n");
         goto end;
     }
 
@@ -516,10 +565,9 @@ end:
     return ret;
 }
 
-int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
+int HAL_AT_TCP_Connect(_IN_ utils_network_pt pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
 {
     int ret = 0;
-    utils_network_pt pNet = (utils_network_pt)pNetwork;
     at_response_t resp = NULL;
     int link_num = 0;
     at_client_t p_client = at_client_get();
@@ -531,9 +579,9 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
         goto end;
     }
 
-    for(link_num = 0; link_num < SIM800C_MAX_TCP_LINK; link_num++)
+    for(link_num = 0; link_num < M26_MAX_TCP_LINK; link_num++)
     {
-        if(sim800c_link[link_num] == eDISCONNECTED)
+        if(m26_link[link_num] == eDISCONNECTED)
         {
             last_tcp_link = link_num;
             break;
@@ -549,7 +597,7 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
             goto end;
         }
         
-        ret = sim800c_init();
+        ret = m26_init();
         if(ret != SUCCESS_RET)
         {
             LOG_ERROR("sim800c init fail!\n");
@@ -564,9 +612,18 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
         goto end;
     }
 
+    /* 解析domain address获取IP */
+    resp->custom_flag = true;
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIDNSGIP=\"%s\"\r", pNetwork->pHostAddress);
+    if(SUCCESS_RET != ret)
+    {
+        LOG_ERROR("build TCP link fail!\n");
+        goto end;
+    }
+
     /* 建立TCP链接 */
     resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPSTART=%d,\"TCP\",\"%s\",\"%d\"\r", link_num, pNet->pHostAddress, pNet->port);
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIOPEN=%d,\"TCP\",\"%s\",\"%d\"\r", link_num, domain_ip_addr, pNetwork->port);
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("build TCP link fail!\n");
@@ -574,14 +631,13 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
     }
     else
     {
-        sim800c_link[link_num] = eCONNECTED;
+        m26_link[link_num] = eCONNECTED;
         /* handle can't be zero */
-        pNet->handle = link_num + 1;
+        pNetwork->handle = link_num + 1;
     }
 end:
     at_delete_resp(resp);
     return ret;
 }
-
