+++ src/at/class/sim800c/sim800c.c
@@ -12,24 +12,21 @@
 * express or implied. See the License for the specific language governing
 * permissions and limitations under the License.
 */
+
 #include <stdio.h>
 #include <string.h>
-#include "stm32f7xx_hal.h"
 #include "at_inf.h"
 #include "utils_net.h"
 #include "at_ringbuff.h"
 #include "at_client.h"
 #include "uiot_import.h"
 
-#define M26_MAX_TCP_LINK        3
-
-extern UART_HandleTypeDef huart2;
-static UART_HandleTypeDef *pAtUart = &huart2;
+#define SIM800C_MAX_TCP_LINK        3
 extern sRingbuff g_ring_buff;    
 extern sRingbuff g_ring_tcp_buff[3];    
-int m26_link[M26_MAX_TCP_LINK] = {0};
+int sim800c_link[SIM800C_MAX_TCP_LINK] = {0};
 extern int last_tcp_link;
-int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_ size_t len)
+int HAL_AT_Read_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
 {
     int ret = 0;
     at_client_t client = at_client_get();
@@ -37,14 +34,8 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
     char temp_char = 0;
     int temp_read_point = client->pRingBuff->readpoint;
     char temp_string[21] = {0};
-    //char temp_tcp = 0;
     int rec_num = 0;
     int link_num = 0;
-
-    if(pAtUart->RxState == HAL_UART_STATE_BUSY_RX)
-    {
-        HAL_SleepMs(10);
-    }
 
     /* clear \r\n */
     ret = at_client_getchar(client, &last_char, GET_RECEIVE_TIMEOUT_MS);
@@ -75,7 +66,7 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
                 if(loop == 8)
                 {   
                     /* divide multi tcp link */
-                    if(0 != strncmp(temp_string, "RECEIVE:", 8))
+                    if(0 != strncmp(temp_string, "RECEIVE,", 8))
                     {
                         client->pRingBuff->readpoint = temp_read_point;
                         break;
@@ -84,12 +75,12 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
                 loop++;
             }while(temp_string[loop-1] != '\n');
             temp_string[loop] = '\0';
-            if(2 == sscanf(temp_string,"RECEIVE: %d, %d:\r\n",&link_num,&rec_num))
+            if(2 == sscanf(temp_string,"RECEIVE,%d,%d:\r\n",&link_num,&rec_num))
             {
                 for(loop = rec_num; loop > 0; loop--)
                 {
-                    ret = at_client_getchar(client, &temp_char, GET_RECEIVE_TIMEOUT_MS);
-                    ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), &temp_char, 1);
+                    ret = at_client_getchar(client, (char *)&temp_char, GET_RECEIVE_TIMEOUT_MS);
+                    ret |= ring_buff_push_data(&(g_ring_tcp_buff[link_num]), (uint8_t *)&temp_char, 1);
                     if(SUCCESS_RET != ret)
                     {
                         LOG_ERROR("copy data to tcp buff fail\n");
@@ -105,21 +96,11 @@ int HAL_AT_Read(_IN_ utils_network_pt pNetwork, _OU_ unsigned char *buffer, _IN_
     }
 
 
-    ret = ring_buff_pop_data(&(g_ring_tcp_buff[pNetwork->handle-1]), buffer, len);
+    ret = HAL_AT_Read(pNetwork, buffer, len);
 
     return ret;
 }
 
-
-int HAL_AT_Write(_IN_ unsigned char *buffer, _IN_ size_t len)
-{   
-    return HAL_UART_Transmit_IT(pAtUart, buffer, len);
-}
-
-int HAL_AT_Read_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
-{
-    return HAL_AT_Read(pNetwork, buffer, len);
-}
 
 int HAL_AT_Write_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer, _IN_ size_t len)
 {
@@ -139,9 +120,9 @@ int HAL_AT_Write_Tcp(_IN_ utils_network_pt pNetwork, _IN_ unsigned char *buffer,
     }
 
     resp->custom_flag = true;
-    at_exec_cmd(resp, at_command, 0, "AT+QISEND=%d,%d\r\n", pNetwork->handle-1, len); 
+    at_exec_cmd(resp, at_command, 0, "AT+CIPSEND=%d,%d\r\n", pNetwork->handle-1, len); 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_data, len, buffer); 
+    ret = at_exec_cmd(resp, at_data, len, (const char *)buffer); 
     HAL_SleepMs(100);
 
     at_delete_resp(resp);
@@ -161,7 +142,7 @@ int HAL_AT_TCP_Disconnect(utils_network_pt pNetwork)
     int ret = SUCCESS_RET;
     
     /* 断开无线链接 */
-    if(m26_link[pNetwork->handle-1] == eCONNECTED)
+    if(sim800c_link[pNetwork->handle-1] == eCONNECTED)
     {
         at_response_t resp = NULL;
         
@@ -173,19 +154,21 @@ int HAL_AT_TCP_Disconnect(utils_network_pt pNetwork)
         }
     
         resp->custom_flag = false;
-        ret = at_exec_cmd(resp, at_command, 0,  "AT+QICLOSE=%d\r",pNetwork->handle-1);
+        ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPCLOSE=%d\r",pNetwork->handle-1);
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("close TCP link fail!\n");
         }
         else
         {
-            m26_link[pNetwork->handle-1] = eDISCONNECTED;
+            sim800c_link[pNetwork->handle-1] = eDISCONNECTED;
         }
         at_delete_resp(resp);
     }
     return ret;
 }
+
+char ip_addr[20] = {0};
 
 static int urc_cpin_recv_judge(const char *data, uint32_t size)
 {
@@ -257,7 +240,7 @@ static int urc_creg_recv_func(const char *data, uint32_t size)
     int stat = 0;
     if(2 == sscanf(data,"+CREG: %d,%d", &n, &stat))
     {
-        if((1 == stat) || (5 == stat))
+        if(1 == stat)
         {
             return SUCCESS_RET;
         }
@@ -267,9 +250,10 @@ static int urc_creg_recv_func(const char *data, uint32_t size)
 }
 
 
-static int urc_cgreg_recv_judge(const char *data, uint32_t size)
+static int urc_cgatt_recv_judge(const char *data, uint32_t size)
 {
-    if(NULL != strstr(data, "+CGREG"))
+    if((data[size-1] == '\n')
+        && (NULL != strstr(data, "+CGATT")))
     {
         return SUCCESS_RET;
     }
@@ -279,14 +263,16 @@ static int urc_cgreg_recv_judge(const char *data, uint32_t size)
     }
 }
 
-static int urc_cgreg_recv_func(const char *data, uint32_t size)
+static int urc_cgatt_recv_func(const char *data, uint32_t size)
 {
-    int n = 0;
-    int stat = 0;
-    if(2 == sscanf(data,"+CGREG: %d,%d",&n, &stat))
+    at_client_t client = at_client_get();
+    int result = 0;
+    if(1 == sscanf(data,"+CGATT: %d",&result))
     {
-        if((1 == stat) || (5 == stat))
+        if(1 == result)
         {
+            at_recv_readline(client);
+            at_recv_readline(client);
             return SUCCESS_RET;
         }
     }
@@ -300,52 +286,11 @@ static int urc_common_recv_func(const char *data, uint32_t size)
     return SUCCESS_RET;
 }
 
-static int urc_qimode_recv_judge(const char *data, uint32_t size)
-{
-    if(NULL != strstr(data, "+QIMODE:"))
-    {
-        return SUCCESS_RET;
-    }
-    else
-    {
-        return FAILURE_RET;
-    }
-}
-
-static int urc_qimode_recv_func(const char *data, uint32_t size)
-{
-    at_client_t client = at_client_get();
-    int n = 0;
-    if(1 == sscanf(data,"+QIMODE: %d\r\n",&n))
-    {
-        if(0 == n)
-        {        
-            at_recv_readline(client);    
-            at_recv_readline(client);
-            return SUCCESS_RET;
-        }
-    }
-    
-    return FAILURE_RET;
-}
-
-static int urc_qideact_recv_judge(const char *data, uint32_t size)
-{
-    if(NULL != strstr(data, "DEACT OK"))
-    {
-        return SUCCESS_RET;
-    }
-    else
-    {
-        return FAILURE_RET;
-    }
-}
-
 static int urc_ip_recv_judge(const char *data, uint32_t size)
 {
     int num1,num2,num3,num4;
     if((data[size-1] == '\n')
-        && (4 == sscanf(data,"%d.%d.%d.%d\r\n",&num1,&num2,&num3,&num4)))
+        && (4 == sscanf(data,"%d.%d.%d.%d",&num1,&num2,&num3,&num4)))
     {   
         if(0<=num1 && num1<=255
          && 0<=num2 && num2<=255
@@ -355,23 +300,13 @@ static int urc_ip_recv_judge(const char *data, uint32_t size)
             return SUCCESS_RET;
         }
     }
-    else
-    {
-        return FAILURE_RET;
-    }
+
+    return FAILURE_RET;   
 }
 
-char dev_ip_addr[20] = {0};
 static int urc_ip_recv_func(const char *data, uint32_t size)
 {
-    strncpy(dev_ip_addr, data, size-2);
-    return SUCCESS_RET;
-}
-
-char domain_ip_addr[20] = {0};
-static int urc_domain_ip_recv_func(const char *data, uint32_t size)
-{
-    strncpy(domain_ip_addr, data, size-2);
+    strncpy(ip_addr, data, size);
     return SUCCESS_RET;
 }
 
@@ -411,24 +346,21 @@ static int urc_close_recv_judge(const char *data, uint32_t size)
     }
 }
 
+
 at_custom custom_table[] = {
     {"AT+CPIN?", 12, urc_cpin_recv_judge, urc_cpin_recv_func},
     {"AT+CSQ", 10, urc_csq_recv_judge, urc_csq_recv_func},
     {"AT+CREG?", 10, urc_creg_recv_judge, urc_creg_recv_func},
-    {"AT+CGREG?", 11, urc_cgreg_recv_judge, urc_cgreg_recv_func},
-    {"AT+QIMODE?", 12, urc_qimode_recv_judge, urc_qimode_recv_func},
-    {"AT+QIDEACT?", 9, urc_qideact_recv_judge, urc_common_recv_func},
-    {"AT+QILOCIP", 13, urc_ip_recv_judge, urc_ip_recv_func},
-    {"AT+QIDNSGIP", 13, urc_ip_recv_judge, urc_domain_ip_recv_func},
-    {"AT+QIDEACT", 10, urc_qideact_recv_judge, urc_common_recv_func},
-    {"AT+QIOPEN", 14, urc_tcp_start_judge, urc_common_recv_func},
-    {"AT+QISEND", 1, urc_send_recv_judge, urc_common_recv_func},
-    {"AT+QICLOSE", 12, urc_close_recv_judge, urc_common_recv_func},
+    {"AT+CGATT?", 9, urc_cgatt_recv_judge, urc_cgatt_recv_func},
+    {"AT+CIFSR", 10, urc_ip_recv_judge, urc_ip_recv_func},
+    {"AT+CIPSTART", 14, urc_tcp_start_judge, urc_common_recv_func},
+    {"AT+CIPSEND", 1, urc_send_recv_judge, urc_common_recv_func},
+    {"AT+CIPCLOSE", 12, urc_close_recv_judge, urc_common_recv_func},
 };
 
 int custom_table_num = sizeof(custom_table) / sizeof(custom_table[0]);
 
-static int m26_init()
+static int sim800c_init()
 {
     int ret = 0;
     at_response_t resp = NULL;
@@ -441,15 +373,12 @@ static int m26_init()
         return FAILURE_RET;
     }
 
-    /* 配置串口接收buf的存储位置 */    
-    HAL_UART_Receive_IT(pAtUart, g_ring_buff.buffer, 1);
-
     //延时等待模块启动
-    HAL_SleepMs(5000);
+    HAL_SleepMs(8000);
 
     /* 去掉串口回显 */
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0, "ATE0\r\n");
+    ret = at_exec_cmd(resp, at_command, 0, "ATE0\r");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("ATE0 set fail!\n");
@@ -465,7 +394,7 @@ static int m26_init()
         
         /* 检查SIM卡状态 */
         resp->custom_flag = true;
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CPIN?\r\n");  
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CPIN?\r");  
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("please check SIM card status!\n");
@@ -474,7 +403,7 @@ static int m26_init()
         
         /* 检查网络强度 */  
         resp->custom_flag = true;
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CSQ\r\n");  
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CSQ\r");  
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("bad signal!\n");
@@ -483,7 +412,7 @@ static int m26_init()
         
         /* 检查网络注册状态 */   
         resp->custom_flag = true;   
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CREG?\r\n"); 
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CREG?\r"); 
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("not network regist!\n");
@@ -492,20 +421,11 @@ static int m26_init()
 
         /*检查GPRS附着状态*/
         resp->custom_flag = true;   
-        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CGREG?\r\n"); 
+        ret |= at_exec_cmd(resp, at_command, 0,  "AT+CGATT?\r"); 
         if(SUCCESS_RET != ret)
         {
             LOG_ERROR("not Attach GPRS!\n");
             continue;
-        }
-
-        /* check TCPIP mode is set */
-        resp->custom_flag = true;
-        ret = at_exec_cmd(resp, at_command, 0,  "AT+QIMODE?\r\n");
-        if(SUCCESS_RET != ret)
-        {
-            LOG_ERROR("check TCPIP mode fail!\n");
-            goto end;
         }
 
         if(SUCCESS_RET == ret)
@@ -521,54 +441,27 @@ static int m26_init()
         goto end;
     }
 
-    /* 设置前台配置 */
-    resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIFGCNT=0\r\n");
-    if(SUCCESS_RET != ret)
-    {
-        LOG_ERROR("set foreground fail!\n");
-        goto end;
-    }
-    
-    /* set the apn */
-    resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QICSGP=1, \"CMNET\"\r\n");
-    if(SUCCESS_RET != ret)
-    {
-        LOG_ERROR("set apn fail!\n");
-        goto end;
-    }
-
-    /* 关闭 GPRS/CSD PDP 场景 */    
-    resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIDEACT\r\n");
-    if(SUCCESS_RET != ret)
-    {
-        LOG_ERROR("shut down GPRS/CSD PDP fail!\n");
-        goto end;
-    }
-
     /* 设置多链路模式 */ 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIMUX=1\r\n");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPMUX=1\r");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("set multi link mode fail!\n");
         goto end;
     }
     
-    /* 启动任务 */    
+    /* 设置APN */    
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIREGAPP\r\n");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CSTT=\"3GNET\"\r");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("set APN fail!\n");
         goto end;
     }
     
-    /* 设备默认响应超时时间为150s，设置为20s */ 
+    /* 建立无线链接 */ 
     resp->custom_flag = false;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIACT\r\n");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIICR\r");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("build wireless link fail!\n");
@@ -577,10 +470,19 @@ static int m26_init()
     
     /* 获取本地IP地址 */  
     resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QILOCIP\r\n");
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIFSR\r");
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("fetch local IP address fail!\n");
+        goto end;
+    }
+
+    /* 快速发送模式 */  
+    resp->custom_flag = false;
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPQSEND=1\r");
+    if(SUCCESS_RET != ret)
+    {
+        LOG_ERROR("fast send mode fail!\n");
         goto end;
     }
 
@@ -589,10 +491,9 @@ end:
     return ret;
 }
 
-int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
+int HAL_AT_TCP_Connect(_IN_ utils_network_pt pNetwork, _IN_ const char *host, _IN_ uint16_t port) 
 {
     int ret = 0;
-    utils_network_pt pNet = (utils_network_pt)pNetwork;
     at_response_t resp = NULL;
     int link_num = 0;
     at_client_t p_client = at_client_get();
@@ -604,9 +505,9 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
         goto end;
     }
 
-    for(link_num = 0; link_num < M26_MAX_TCP_LINK; link_num++)
+    for(link_num = 0; link_num < SIM800C_MAX_TCP_LINK; link_num++)
     {
-        if(m26_link[link_num] == eDISCONNECTED)
+        if(sim800c_link[link_num] == eDISCONNECTED)
         {
             last_tcp_link = link_num;
             break;
@@ -622,7 +523,7 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
             goto end;
         }
         
-        ret = m26_init();
+        ret = sim800c_init();
         if(ret != SUCCESS_RET)
         {
             LOG_ERROR("sim800c init fail!\n");
@@ -637,18 +538,9 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
         goto end;
     }
 
-    /* 解析domain address获取IP */
-    resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIDNSGIP=\"%s\"\r", pNet->pHostAddress);
-    if(SUCCESS_RET != ret)
-    {
-        LOG_ERROR("build TCP link fail!\n");
-        goto end;
-    }
-
     /* 建立TCP链接 */
     resp->custom_flag = true;
-    ret = at_exec_cmd(resp, at_command, 0,  "AT+QIOPEN=%d,\"TCP\",\"%s\",\"%d\"\r", link_num, domain_ip_addr, pNet->port);
+    ret = at_exec_cmd(resp, at_command, 0,  "AT+CIPSTART=%d,\"TCP\",\"%s\",\"%d\"\r", link_num, pNetwork->pHostAddress, pNetwork->port);
     if(SUCCESS_RET != ret)
     {
         LOG_ERROR("build TCP link fail!\n");
@@ -656,13 +548,14 @@ int HAL_AT_TCP_Connect(_IN_ void * pNetwork, _IN_ const char *host, _IN_ uint16_
     }
     else
     {
-        m26_link[link_num] = eCONNECTED;
+        sim800c_link[link_num] = eCONNECTED;
         /* handle can't be zero */
-        pNet->handle = link_num + 1;
+        pNetwork->handle = link_num + 1;
     }
 end:
     at_delete_resp(resp);
     return ret;
 }
+
