+++ src/utils/utils_net.c
@@ -80,55 +80,6 @@ static int connect_ssl(utils_network_pt pNetwork)
     }
 }
 
-/*** TCP connection ***/
-static int read_tcp(utils_network_pt pNetwork, unsigned char *buffer, size_t len, uint32_t timeout_ms)
-{
-    if (NULL == pNetwork) {
-        LOG_ERROR("network is null");
-        return FAILURE_RET;
-    }
-
-    return HAL_TCP_Read((uintptr_t)pNetwork->handle, buffer, len, timeout_ms);
-}
-
-
-static int write_tcp(utils_network_pt pNetwork, unsigned char *buffer, size_t len, uint32_t timeout_ms)
-{
-    if (NULL == pNetwork) {
-        LOG_ERROR("network is null");
-        return FAILURE_RET;
-    }
-
-    return HAL_TCP_Write((uintptr_t)pNetwork->handle, buffer, len, timeout_ms);
-}
-
-static int disconnect_tcp(utils_network_pt pNetwork)
-{
-    if (NULL == pNetwork) {
-        LOG_ERROR("network is null");
-        return FAILURE_RET;
-    }
-
-    HAL_TCP_Disconnect(pNetwork->handle);
-    pNetwork->handle = (uintptr_t)(-1);
-    return SUCCESS_RET;
-}
-
-static int connect_tcp(utils_network_pt pNetwork)
-{
-    if (NULL == pNetwork) {
-        LOG_ERROR("network is null");
-        return FAILURE_RET;
-    }
-
-    pNetwork->handle = HAL_TCP_Connect(pNetwork->pHostAddress, pNetwork->port);
-    if (pNetwork->handle == (uintptr_t)(-1)) {
-        return FAILURE_RET;
-    }
-
-    return SUCCESS_RET;
-}
-
 #elif SUPPORT_AT_CMD
 /* connect TCP by AT cmd through uart */
 static int at_read_tcp(utils_network_pt pNetwork, unsigned char *buffer, size_t len, uint32_t timeout_ms)
@@ -227,19 +178,11 @@ int utils_net_read(utils_network_pt pNetwork, unsigned char *buffer, size_t len,
 {
     int ret = 0;
 #ifdef SUPPORT_TLS
-    if (NULL != pNetwork->ca_crt) {
         ret = read_ssl(pNetwork, buffer, len, timeout_ms);
-    }
-    else
-    {
-        ret = read_tcp(pNetwork, buffer, len, timeout_ms);
-    }
 #elif SUPPORT_AT_CMD
         ret = at_read_tcp(pNetwork, buffer, len, timeout_ms);
 #else
-    if (NULL == pNetwork->ca_crt) {
         ret = read_tcp(pNetwork, buffer, len, timeout_ms);
-    }   
 #endif
 
     return ret;
@@ -249,19 +192,11 @@ int utils_net_write(utils_network_pt pNetwork,unsigned char *buffer, size_t len,
 {
     int ret = 0;
 #ifdef SUPPORT_TLS
-    if (NULL != pNetwork->ca_crt) {
         ret = write_ssl(pNetwork, buffer, len, timeout_ms);
-    }
-    else
-    {
-        ret = write_tcp(pNetwork, buffer, len, timeout_ms);
-    }
 #elif SUPPORT_AT_CMD
         ret = at_write_tcp(pNetwork, buffer, len, timeout_ms);
 #else
-    if (NULL == pNetwork->ca_crt) {
         ret = write_tcp(pNetwork, buffer, len, timeout_ms);
-    }
 #endif
 
     return ret;
@@ -271,19 +206,11 @@ int utils_net_disconnect(utils_network_pt pNetwork)
 {
     int ret = 0;
 #ifdef SUPPORT_TLS
-    if (NULL != pNetwork->ca_crt) {
         ret = disconnect_ssl(pNetwork);
-    }
-    else
-    {
-        ret = disconnect_tcp(pNetwork);
-    }
 #elif SUPPORT_AT_CMD
         ret = at_disconnect_tcp(pNetwork);
 #else
-    if (NULL == pNetwork->ca_crt) {
         ret = disconnect_tcp(pNetwork);
-    }
 #endif
 
     return  ret;
@@ -293,19 +220,11 @@ int utils_net_connect(utils_network_pt pNetwork)
 {
     int ret = 0;
 #ifdef SUPPORT_TLS
-    if (NULL != pNetwork->ca_crt) {
         ret = connect_ssl(pNetwork);
-    }
-    else
-    {
-        ret = connect_tcp(pNetwork);
-    }
 #elif SUPPORT_AT_CMD
         ret = at_connect_tcp(pNetwork);
 #else
-    if (NULL == pNetwork->ca_crt) {
         ret = connect_tcp(pNetwork);
-    }
 #endif
 
     return ret;
@@ -319,7 +238,6 @@ int utils_net_init(utils_network_pt pNetwork, const char *host, uint16_t port, u
     }
     pNetwork->pHostAddress = host;
     pNetwork->port = port;
-#ifndef SUPPORT_AT_CMD
     pNetwork->authmode = authmode;
     pNetwork->ca_crt = ca_crt;
 
@@ -328,7 +246,6 @@ int utils_net_init(utils_network_pt pNetwork, const char *host, uint16_t port, u
     } else {
         pNetwork->ca_crt_len = strlen(ca_crt);
     }
-#endif 
 
     pNetwork->handle = 0;
     pNetwork->read = utils_net_read;
