+++ tools/scripts/default_settings.mk
@@ -1,6 +1,7 @@
 PLATFORM_CC            ?= gcc
 PLATFORM_AR            ?= ar
 PLATFORM_OS            ?= linux
+PLATFORM_MODULE		   ?= esp8266
 
 FEATURE_MQTT_COMM_ENABLED           ?= y
 FEATURE_DEVICE_SHADOW_ENABLED       ?= n
@@ -9,6 +10,7 @@ FEATURE_DEVICE_MODEL_ENABLED        ?= y
 FEATURE_FILE_UPLOAD_ENABLED         ?= y
 FEATURE_AUTH_MODE_DYNAMIC           ?= n
 FEATURE_SUPPORT_TLS					?= n
+FEATURE_SUPPORT_AT_CMD				?= n
 
 FEATURE_SDK_TESTS_ENABLED           ?= n    #是否开启单元测试编译	仅支持linux
 FEATURE_MQTT_RMDUP_MSG_ENABLED      ?= n    #是否开启MQTT消息去重能力
