+++ tools/scripts/parse_make_settings.mk
@@ -5,6 +5,7 @@ SETTING_VARS := \
     PLATFORM_CC \
     PLATFORM_AR \
     PLATFORM_OS \
+    PLATFORM_MODULE \
 
 SWITCH_VARS := \
     FEATURE_MQTT_COMM_ENABLED \
@@ -15,6 +16,7 @@ SWITCH_VARS := \
 	FEATURE_AUTH_MODE_DYNAMIC \
 	FEATURE_DEVICE_MODEL_ENABLED \
 	FEATURE_FILE_UPLOAD_ENABLED	\
+    FEATURE_SUPPORT_AT_CMD
 
 $(foreach v, \
     $(SETTING_VARS) $(SWITCH_VARS), \
